实体存在性加固(三层防线) + 身份兜底 + 灭在多空指令
- 移除加固三层防线：入口抢先 + 事件反否决 + 结构自愈(不依赖官方加入入口)
- 身份兜底：getUUID/getStringUUID空值兜底 + 身份快照拒绝空值 + volatile
- 自然清除防护：removeWhenFarAway/isPersistenceRequired固定值
- 灭在多空命令 /yiz sx mzdk + 玩家近战触发百分比伤害

+++ src/main/java/net/minecraft/client/yiz/xian/entity/ShanlinshouzheEntity.java
@@ -196,7 +196,7 @@ public class ShanlinshouzheEntity extends YizxianMob {
     @Override
     public boolean doHurtTarget(Entity target) {
         if (target instanceof LivingEntity lt) {
-            float dream = (float) this.getAttributeValue(YizAttributes.FIRST_DREAM.get());
+            double dream = this.getAttributeValue(YizAttributes.FIRST_DREAM.get());
             if (dream > 0) {
                 net.minecraft.client.yiz.tool.health.EntityASMUtil.applyDreamDamage(this, lt, dream);
             }
