+++ src/main/java/net/minecraft/client/yiz/xian/entity/TaxutiEntity.java
@@ -230,7 +230,7 @@ public class TaxutiEntity extends YizxianMob implements GeoEntity {
     @Override
     public boolean doHurtTarget(Entity target) {
         if (target instanceof LivingEntity lt) {
-            float dream = (float) this.getAttributeValue(YizAttributes.FIRST_DREAM.get());
+            double dream = this.getAttributeValue(YizAttributes.FIRST_DREAM.get());
             if (dream > 0) {
                 net.minecraft.client.yiz.tool.health.EntityASMUtil.applyDreamDamage(this, lt, dream);
             }
